break up file and fix an eslint pronto warning
