+++ .pronto_eslint_npm.yml
@@ -1,3 +1,3 @@
 # .pronto_eslint_npm.yml
 eslint_executable: './node_modules/.bin/eslint'
-files_to_lint: 'app/javascript/**/*'
+files_to_lint: 'app/javascript/*/*'
