+++ app/searchers/query_grammar/transformer.rb
@@ -1,0 +1,4 @@
+module QueryGrammar
+  class Transformer < Parslet::Transform
+  end
+end
