+++ spec/searchers/query_grammar/transformer_spec.rb
@@ -1,0 +1,3 @@
+RSpec.describe QueryGrammar::Transformer do
+  skip "Nothing to test here yet"
+end
